# Speed up Info\Error\etc extensions by ~10-20 percent. 