--- a/Vostok.Logging.Abstractions.Tests/Vostok.Logging.Abstractions.Tests.csproj
+++ b/Vostok.Logging.Abstractions.Tests/Vostok.Logging.Abstractions.Tests.csproj
@@ -9,6 +9,7 @@
   <ItemGroup>
     <PackageReference Include="FluentAssertions" Version="5.4.1" />
     <PackageReference Include="NSubstitute" Version="3.1.0" />
+    <PackageReference Include="BenchmarkDotNet" Version="0.13.0" />
   </ItemGroup>
 
   <ItemGroup>