--- a/Vostok.Logging.Abstractions/Extensions/LogEventExtensions.cs
+++ b/Vostok.Logging.Abstractions/Extensions/LogEventExtensions.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Runtime.CompilerServices;
 using JetBrains.Annotations;
+using Vostok.Commons.Collections;
 using Vostok.Logging.Abstractions.Helpers;
 using Vostok.Logging.Abstractions.Values;
 
@@ -10,6 +11,8 @@ namespace Vostok.Logging.Abstractions
     [PublicAPI]
     public static class LogEventExtensions
     {
+        private static readonly Func<string, bool> IsPositionalNameCachedFunction = IsPositionalName;
+
         [Pure]
         public static LogEvent WithObjectProperties<T>(this LogEvent @event, T @object, bool allowOverwrite = true, bool allowNullValues = true)
             => @event.MutateProperties(eventProperties => eventProperties.WithObjectProperties(@object, allowOverwrite, allowNullValues));
@@ -24,26 +27,21 @@ namespace Vostok.Logging.Abstractions
                 properties =>
                 {
                     var templatePropertyNames = TemplatePropertiesExtractor.ExtractPropertyNames(@event.MessageTemplate);
-
-                    if (ShouldInferNamesForPositionalParameters(templatePropertyNames))
-                    {
-                        // (iloktionov): Name positional parameters with corresponding placeholder names from template:
-                        for (var i = 0; i < Math.Min(parameters.Length, templatePropertyNames.Length); i++)
-                            properties = properties.Set(templatePropertyNames[i], parameters[i]);
-
-                        if (parameters.Length > templatePropertyNames.Length)
-                            for (var i = templatePropertyNames.Length; i < parameters.Length; i++)
-                                properties = properties.Set(i.ToString(), parameters[i]);
-                    }
-                    else
-                    {
-                        // (iloktionov): Name positional parameters with their indices:
-                        for (var i = 0; i < parameters.Length; i++)
-                            properties = properties.Set(i.ToString(), parameters[i]);
-                    }
-
-                    return properties;
+                    return FillImmutableArrayDictionaryWithProperties(templatePropertyNames, parameters, properties);
                 });
+        }
+
+        [Pure]
+        internal static ImmutableArrayDictionary<string, object> GenerateInitialParameters(string messageTemplate, object[] parameters)
+        {
+            if (parameters == null || parameters.Length == 0)
+                return null;
+
+            var properties = parameters.Length > 4 ? LogEvent.CreateProperties(parameters.Length) : LogEvent.CreateProperties();
+
+            var templatePropertyNames = TemplatePropertiesExtractor.ExtractPropertyNames(messageTemplate);
+
+            return FillImmutableArrayDictionaryWithProperties(templatePropertyNames, parameters, properties);
         }
 
         internal static bool HasMatchingSourceContexts(this LogEvent @event, string[] contexts)
@@ -65,12 +63,34 @@ namespace Vostok.Logging.Abstractions
                     sourceContext.Any(value => value.StartsWith(context, StringComparison.OrdinalIgnoreCase)));
         }
 
+        private static ImmutableArrayDictionary<string, object> FillImmutableArrayDictionaryWithProperties(string[] templatePropertyNames, object[] parameters, ImmutableArrayDictionary<string, object> properties)
+        {
+            if (ShouldInferNamesForPositionalParameters(templatePropertyNames))
+            {
+                // (iloktionov): Name positional parameters with corresponding placeholder names from template:
+                for (var i = 0; i < Math.Min(parameters.Length, templatePropertyNames.Length); i++)
+                    properties = properties.Set(templatePropertyNames[i], parameters[i]);
+
+                if (parameters.Length > templatePropertyNames.Length)
+                    for (var i = templatePropertyNames.Length; i < parameters.Length; i++)
+                        properties = properties.Set(i.ToString(), parameters[i]);
+            }
+            else
+            {
+                // (iloktionov): Name positional parameters with their indices:
+                for (var i = 0; i < parameters.Length; i++)
+                    properties = properties.Set(i.ToString(), parameters[i]);
+            }
+
+            return properties;
+        }
+
         private static bool ShouldInferNamesForPositionalParameters(string[] propertyNames)
         {
             if (propertyNames.Length == 0)
                 return false;
 
-            if (propertyNames.All(IsPositionalName))
+            if (propertyNames.All(IsPositionalNameCachedFunction))
                 return false;
 
             return true;