In one of my apps, log.Info is the second major consumer of time and CPU and the fifth spammer for memory traffic.
Also I have a significant number of calls to Debug and Warn extensions.

So a little refactoring to avoid unnecessary closures and delegates, as well as a little savings in rebuilding the LogEvent, speed up these extensions by up to 10-20 percent (depending on the input). And it slightly reduces the traffic of objects.

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1415 (21H1/May2021Update)
Intel Core i7-4771 CPU 3.50GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.101
  [Host]      : .NET 5.0.13 (5.0.1321.56516), X64 RyuJIT


|                Method |     Mean |  Gen 0 | Allocated |
|---------------------- |---------:|-------:|----------:|
|     WithTwoParameters | 356.7 ns | 0.0896 |     376 B |
|  WithTwoParametersOld | 413.9 ns | 0.1431 |     600 B |


|                Method |     Mean |  Gen 0 | Allocated |
|---------------------- |---------:|-------:|----------:|
|    WithFourParameters | 488.4 ns | 0.1125 |     472 B |
| WithFourParametersOld | 544.4 ns | 0.1659 |     696 B |


|                Method |     Mean |  Gen 0 | Allocated |
|---------------------- |---------:|-------:|----------:|
|    WithFiveParameters | 551.4 ns | 0.1297 |     544 B |
| WithFiveParametersOld | 648.6 ns | 0.2289 |     960 B |